evm: add negative tests

+++ evm/src/MatchingEngine/assets/Errors.sol
@@ -12,7 +12,7 @@ error ErrInvalidTargetRouter(uint16 chain);
 
 error ErrInvalidSourceRouter(bytes32 sender, bytes32 expected);
 
-error ErrBidPriceTooHigh(uint128 bidPrice, uint256 maxPrice);
+error ErrBidPriceTooHigh(uint128 bidPrice, uint128 maxPrice);
 
 error ErrAuctionPeriodExpired();
